Developing Day 1
Started developing. Made some changes to the UI and upgrade so that they can scale and also created them, made another SO I think and then also created the  GameManager, or worked on it. Also realized should have used decimal and nomt BigInteger, so thne changed that, hopefully numbers do not get that big too quickly.