Developing Day 4
Added the text so that it can display the variables, that is working well and changed all the regular variables to also variable classes as that is also mroe convenient. Still need to work on the upgrades and the buy function, but most of it seems to be going well.