Developing Day 7
Added the 2 upgrades in and then yeah it worked pretty well! Adding the first a one was no bugs actually except for one where it was displaying the wrong text on the UI. The b upgrade was a bit trickier as I had to fix some stuff with like the custom upgrade initializer and the purchase function. Other than that it was working very well!